requirements.txt
Babel==2.6.0
boto==2.49.0
Django==2.1.11
django-cors-headers==3.1.0
django-storages==1.7.1
djangorestframework==3.9.2
django-heroku==0.3.1
gunicorn==19.9.0
html5lib==1.0.1
whitenoise==4.1.3
Pillow==5.3.0
pipreqs==0.4.9